Move MIME type checking to its own module
